rebit funciona
rabit funcionando con inventario de forma baga hay que adapatar el inventario para que coincida con algunos parametros extra como detalle y cantidad en unidades de medida

+++ services/inventario/src/main/java/com/donaton/inventario/model/Inventario.java
@@ -18,5 +18,5 @@ public class Inventario {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String producto;
-    private Integer stock;
+    private Double stock;
 }
